code/requirements.txt
tensorflow
opencv-python
streamlit
deepstar-sdk
numpy
